refactor: clean up stories and rename package
- Rename package from `ui-modal` to `ui-dialog`
- Create a custom component to eliminate duplicate code in stories
